requirements.txt
attrs==22.2.0
autopep8==2.0.1
aws-cdk-lib==2.60.0
aws-cdk.asset-awscli-v1==2.2.52
aws-cdk.asset-kubectl-v20==2.1.1
aws-cdk.asset-node-proxy-agent-v5==2.0.42
cattrs==22.2.0
clg==1.1.1
constructs==10.1.235
exceptiongroup==1.1.0
jsii==1.74.0
loadconfig==0.0.9
publication==0.0.3
pycodestyle==2.10.0
python-dateutil==2.8.2
python-dotenv==0.21.1
PyYAML==3.11
six==1.9.0
tomli==2.0.1
typeguard==2.13.3
typing_extensions==4.4.0
